Consolidated all panel code in one file, and removed all external dependencies. ITS A LIBRARY NOW (still needs refinement)
