+++ src/animation/foldhaus_animation_interface.h
@@ -1,11 +1,3 @@
-// TODO
-// [] - Moving animation blocks
-// [] - dragging beginning and end of time blocks
-// [] - creating a timeblock with a specific animation
-// [x] - play, pause, stop, 
-// [] - setting the start and end of the animation system
-// [] - displaying multiple layers
-// [] - 
 
 FOLDHAUS_INPUT_COMMAND_PROC(DeleteAnimationBlock)
 {
